+++ MyRadio/Infrastructure/Services/NetworkService.swift
@@ -1,0 +1,13 @@
+//
+//  NetworkService.swift
+//  MyRadio
+//
+//  Created by Aleksei on 22.06.2023.
+//
+
+import Foundation
+
+final class NetworkService {
+    
+    init() {}
+}
