+++ MyRadio/Presentation/Controllers/Settings/Cells/CollectionCell/AppStyleCell.swift
