+++ MyRadio/Resources/Colors.swift
@@ -8,6 +8,8 @@
 import UIKit
 
 enum Colors {
+    // MARK: - Classic
+    
     enum ClassicColorSet {
         static var acceptColor = UIColor { trait in
             if trait.userInterfaceStyle == .dark {
@@ -52,6 +54,8 @@ enum Colors {
         }
     }
     
+    // MARK: - Custom
+    
     enum CustomColorSet {
         static var acceptColor = UIColor { trait in
             if trait.userInterfaceStyle == .dark {
@@ -94,6 +98,8 @@ enum Colors {
         }
     }
     
+    // MARK: - Ocean
+    
     enum OceanColorSet {
         static var acceptColor = UIColor { trait in
             if trait.userInterfaceStyle == .dark {
@@ -134,44 +140,5 @@ enum Colors {
                 return UIColor(hexString: "#F2F2F7")
             }
         }
-    }
-}
-
-struct ColorSet {
-    let title: String
-    var acceptColor: UIColor
-    var tintColor: UIColor
-    var secondColor: UIColor
-    var unselectedColor: UIColor
-    var mainColor: UIColor
-    var backgroundColor: UIColor
-    
-    static func getColorSetCollection() -> [ColorSet] {
-        return [
-            ColorSet(
-                title: Localizable.ColorSetTitle.system,
-                acceptColor: Colors.ClassicColorSet.acceptColor,
-                tintColor: .systemGreen,
-                secondColor: Colors.ClassicColorSet.secondColor,
-                unselectedColor: Colors.ClassicColorSet.unselectedColor,
-                mainColor: Colors.ClassicColorSet.mainColor,
-                backgroundColor: Colors.ClassicColorSet.backgroundColor),
-            ColorSet(
-                title: Localizable.ColorSetTitle.custom,
-                acceptColor: Colors.CustomColorSet.acceptColor,
-                tintColor: Colors.CustomColorSet.acceptColor,
-                secondColor: Colors.CustomColorSet.secondColor,
-                unselectedColor: Colors.CustomColorSet.unselectedColor,
-                mainColor: Colors.CustomColorSet.mainColor,
-                backgroundColor: Colors.CustomColorSet.backgroundColor),
-            ColorSet(
-                title: Localizable.ColorSetTitle.ocean,
-                acceptColor: Colors.OceanColorSet.acceptColor,
-                tintColor: Colors.CustomColorSet.acceptColor,
-                secondColor: Colors.OceanColorSet.secondColor,
-                unselectedColor: Colors.OceanColorSet.unselectedColor,
-                mainColor: Colors.OceanColorSet.mainColor,
-                backgroundColor: Colors.OceanColorSet.backgroundColor),
-        ]
     }
 }
